beautifications, led brightness improvements, makefile setup
