+++ Makefile
@@ -11,7 +11,7 @@ subdirs:
 	$(MAKE) -C testapp
 	$(MAKE) -C musicapp
 	$(MAKE) -C rfapp
-	$(MAKE) -C l0ungel1cht
+	$(MAKE) -C l0unge
 	$(MAKE) -C hackrfapp
 	$(MAKE) -C bootloader
 	$(MAKE) -C flashapp
@@ -43,7 +43,7 @@ clean:
 	$(MAKE) -C testapp clean
 	$(MAKE) -C musicapp clean
 	$(MAKE) -C rfapp clean
-	$(MAKE) -C l0ungel1cht clean
+	$(MAKE) -C l0unge clean
 	$(MAKE) -C bootloader clean
 	$(MAKE) -C flashapp clean
 	$(MAKE) -C l0dables clean
